user/ admin roles

+++ routes/web.php
@@ -15,28 +15,24 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+
+// Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 Auth::routes();
 
-Route::get('/restaurant', [RestaurantController::class, 'index'])->name('restaurant.index');
-Route::get('/restaurant/show/{restaurant}', [RestaurantController::class, 'show'])->name('restaurant.show');
-Route::get('/restaurant/edit/{restaurant}', [RestaurantController::class, 'edit'])->name('restaurant.edit');
-Route::get('/restaurant/create', [RestaurantController::class, 'create'])->name('restaurant.create');
-Route::post('/restaurant/store', [RestaurantController::class, 'store'])->name('restaurant.store');
-Route::put('/restaurant/update/{restaurant}', [RestaurantController::class, 'update'])->name('restaurant.update');
-Route::delete('/restaurant/delete/{restaurant}', [RestaurantController::class, 'destroy'])->name('restaurant.delete');
+Route::get('/restaurant', [RestaurantController::class, 'index'])->name('restaurant.index')->middleware('gate:user');
+Route::get('/restaurant/show/{restaurant}', [RestaurantController::class, 'show'])->name('restaurant.show')->middleware('gate:user');
+Route::put('/restaurant/rate/{menu}', [RestaurantController::class, 'rateMenu'])->name('restaurant.rate')->middleware('gate:user');
+Route::get('/restaurant/edit/{restaurant}', [RestaurantController::class, 'edit'])->name('restaurant.edit')->middleware('gate:admin');
+Route::get('/restaurant/create', [RestaurantController::class, 'create'])->name('restaurant.create')->middleware('gate:admin');
+Route::post('/restaurant/store', [RestaurantController::class, 'store'])->name('restaurant.store')->middleware('gate:admin');
+Route::put('/restaurant/update/{restaurant}', [RestaurantController::class, 'update'])->name('restaurant.update')->middleware('gate:admin');
+Route::delete('/restaurant/delete/{restaurant}', [RestaurantController::class, 'destroy'])->name('restaurant.delete')->middleware('gate:admin');
 
-Route::get('/menu', [MenuController::class, 'index'])->name('menu.index');
-Route::get('/menu/show/{menu}', [MenuController::class, 'show'])->name('menu.show');
-Route::get('/menu/edit/{menu}', [MenuController::class, 'edit'])->name('menu.edit');
-Route::get('/menu/create', [MenuController::class, 'create'])->name('menu.create');
-Route::post('/menu/store', [MenuController::class, 'store'])->name('menu.store');
-Route::put('/menu/update/{menu}', [MenuController::class, 'update'])->name('menu.update');
-Route::delete('/menu/delete/{menu}', [MenuController::class, 'destroy'])->name('menu.delete');
-
-Auth::routes();
-
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/menu', [MenuController::class, 'index'])->name('menu.index')->middleware('gate:admin');
+Route::get('/menu/show/{menu}', [MenuController::class, 'show'])->name('menu.show')->middleware('gate:admin');
+Route::get('/menu/edit/{menu}', [MenuController::class, 'edit'])->name('menu.edit')->middleware('gate:admin');
+Route::get('/menu/create', [MenuController::class, 'create'])->name('menu.create')->middleware('gate:admin');
+Route::post('/menu/store', [MenuController::class, 'store'])->name('menu.store')->middleware('gate:admin');
+Route::put('/menu/update/{menu}', [MenuController::class, 'update'])->name('menu.update')->middleware('gate:admin');
+Route::delete('/menu/delete/{menu}', [MenuController::class, 'destroy'])->name('menu.delete')->middleware('gate:admin');
